+++ projects/stage-0/user-platform/user-web/src/main/java/org/geektimes/projects/user/orm/jpa/DelegatingEntityManager.java
@@ -1,6 +1,6 @@
 package org.geektimes.projects.user.orm.jpa;
 
-import org.geektimes.context.ComponentContext;
+import org.geektimes.context.JndiComponentContext;
 
 import javax.annotation.PostConstruct;
 import javax.persistence.*;
@@ -58,7 +58,7 @@ public class DelegatingEntityManager implements EntityManager {
             throw new RuntimeException(e);
         }
         // 增加 JNDI 引用处理
-        ComponentContext componentContext = ComponentContext.getInstance();
+        JndiComponentContext componentContext = JndiComponentContext.getInstance();
 
         for (String propertyName : properties.stringPropertyNames()) {
             String propertyValue = properties.getProperty(propertyName);
